requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Configuration & Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0
pydantic[email]>=2.10.0

# YAML Processing
pyyaml>=6.0.2

# HTTP Client
httpx>=0.28.1

# Security - JWT Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20

# Rate Limiting
slowapi>=0.1.9

# Caching
redis>=5.2.0
aioredis>=2.0.2

# Logging & Observability
structlog>=24.4.0
prometheus-client>=0.21.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
httpx>=0.28.1
fakeredis>=2.26.0

# Code Quality
ruff>=0.8.0
